Fix functional tests
Finally fix the existing functional tests. Note that we still need to refactor a lot and create new tests

+++ app/Informulate/Users/ProfileRepository.php
@@ -59,27 +59,29 @@ class ProfileRepository
 	 */
 	public function updateImage(Profile $profile)
 	{
-		// Upload the file to a temp folder and get the file extension
-		$upload = Input::file('image');
-		$extension = $upload->getClientOriginalExtension();
+		if (Input::has('image')) {
+			// Upload the file to a temp folder and get the file extension
+			$upload = Input::file('image');
+			$extension = $upload->getClientOriginalExtension();
 
-		// Use intervention image to create the new file with the proper dimensions
-		// TODO: Instead of resize, figure out a way to maintain the aspect ratio
-		$image = Image::make($upload);//->resize(150, 150);
-		// Fix the new file extension
-		$image->extension = $extension;
-		// Prepare to move the file to the proper folder and with the proper unique name
-		$newPath = public_path() . DIRECTORY_SEPARATOR . 'images' . DIRECTORY_SEPARATOR . 'upload';
-		$newName = $profile->id . '_t.' . $image->extension;
+			// Use intervention image to create the new file with the proper dimensions
+			// TODO: Instead of resize, figure out a way to maintain the aspect ratio
+			$image = Image::make($upload);//->resize(150, 150);
+			// Fix the new file extension
+			$image->extension = $extension;
+			// Prepare to move the file to the proper folder and with the proper unique name
+			$newPath = public_path() . DIRECTORY_SEPARATOR . 'images' . DIRECTORY_SEPARATOR . 'upload';
+			$newName = $profile->id . '_t.' . $image->extension;
 
-		// Remove the existing file if needed
-		if (!empty($profile->image)) {
-			unlink(public_path() . DIRECTORY_SEPARATOR . 'images' . DIRECTORY_SEPARATOR . 'upload' . DIRECTORY_SEPARATOR . $profile->image);
+			// Remove the existing file if needed
+			if (!empty($profile->image)) {
+				unlink(public_path() . DIRECTORY_SEPARATOR . 'images' . DIRECTORY_SEPARATOR . 'upload' . DIRECTORY_SEPARATOR . $profile->image);
+			}
+
+			// Save the image and the profile
+			$image->save($newPath . '/' . $newName);
+			$profile->image = $newName;
+			$profile->save();
 		}
-
-		// Save the image and the profile
-		$image->save($newPath . '/' . $newName);
-		$profile->image = $newName;
-		$profile->save();
 	}
 }
